Feature: Texture Atlas (#15)
Adding a `dgm::TextureAtlas` that can accumulate multiple textures into
a single, big one. This makes it easy to put multiple disparate entities
into a single VBO and optimize drawcalls needed to render them.

Each texture is added either as tileset (with associated `dgm::Clip`) or
as spritesheet (with associated `dgm::AnimationStates`). This makes it
easy to figure out how the texture was offset and to migrate existing
code with zero to no effort.